Feat: 지도 시작 위치 연남동 고정, 검색 결과 아이콘 변경, 맛집 상세 바텀시트 버그 수정 (#14)

+++ app/src/main/java/com/example/tastezzip/model/response/cafeteria/detail/Video.kt
@@ -8,7 +8,7 @@ data class Video(
     @SerializedName("platform")
     val platform: String = "",
     @SerializedName("starAverage")
-    val starAverage: Int = 0,
+    val starAverage: Double = 0.0,
     @SerializedName("status")
     val status: String = "",
     @SerializedName("thumbnailUrl")
